client-agent/requirements.txt
requests>=2.31.0
psutil>=5.9.8
mss>=9.0.1
Pillow>=10.3.0
python-dotenv>=1.0.1
pynput>=1.7.6

